requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flasgger==0.9.7.1
gunicorn==23.0.0
pytest==8.3.5
pytest-cov==6.1.1
